requirements.txt
Django==2.2
djangorestframework==3.10.3
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
